requirements.txt
flask==3.0.2
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.118.0
google-cloud-firestore==2.14.0
python-dotenv==1.0.1
requests==2.31.0
playwright==1.42.0
langchain==0.1.12
langchain-openai==0.0.8
openai==1.14.1
beautifulsoup4==4.12.3
aiohttp==3.9.3
asyncio==3.4.3
werkzeug==3.0.1
pytest==8.0.2
pytest-asyncio==0.23.5
pytest-playwright==0.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-timeout==2.2.0
pytest-xdist==3.5.0
pytest-rerunfailures==14.0
pytest-sugar==1.0.0
pytest-html==4.1.1
pytest-metadata==3.1.0
pytest-selenium==4.0.0
pytest-bdd==7.0.1
pytest-dependency==0.5.1
pytest-env==1.0.1
pytest-flask==1.3.0 